Initial move to v1.2 which will mirror green v1.2
