+++ src/main/java/com/artistech/ee/beans/Data.java
@@ -1,8 +1,9 @@
 /*
  * Copyright 2017 ArtisTech, Inc.
  */
-package com.artistech.ee.web;
+package com.artistech.ee.beans;
 
+import com.artistech.utils.ExternalProcess;
 import java.io.File;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
