+++ src/main/java/com/artistech/ee/web/ProcessOutput.java
@@ -3,6 +3,8 @@
  */
 package com.artistech.ee.web;
 
+import com.artistech.ee.beans.Data;
+import com.artistech.ee.beans.DataManager;
 import java.io.IOException;
 import java.io.PrintWriter;
 import javax.servlet.ServletException;
