+++ src/main/java/com/artistech/utils/ExternalProcess.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2017 ArtisTech, Inc.
  */
-package com.artistech.ee.web;
+package com.artistech.utils;
 
 /**
  *
