+++ src/main/java/com/artistech/utils/Mailbox.java
@@ -4,7 +4,7 @@
  * http://creativecommons.org/licenses/by/3.0/ or send a letter to Creative
  * Commons, 171 Second Street, Suite 300, San Francisco, California, 94105, USA.
  */
-package com.artistech.ee.web;
+package com.artistech.utils;
 
 import java.util.ArrayList;
