+++ src/main/webapp/WEB-INF/web.xml
@@ -6,7 +6,7 @@
     </servlet>
     <servlet>
         <servlet-name>CAMR</servlet-name>
-        <servlet-class>com.artistech.ee.web.CAMR</servlet-class>
+        <servlet-class>com.artistech.ee.yellow.CAMR</servlet-class>
         <init-param>
             <param-name>path</param-name>
             <param-value>/home/matta/Dev/camr</param-value>
@@ -18,7 +18,7 @@
     </servlet>
     <servlet>
         <servlet-name>LiberalEvent</servlet-name>
-        <servlet-class>com.artistech.ee.web.LiberalEvent</servlet-class>
+        <servlet-class>com.artistech.ee.yellow.LiberalEvent</servlet-class>
         <init-param>
             <param-name>path</param-name>
             <param-value>/work/Documents/FOUO/EntityExtraction/LiberalEvent</param-value>
